feat: Initiator JS (#74)
* initial commit for sdk_js

* js wrapper callback approach

* fix test.js

* Add cohort mock, rebase onto recent master

* Add cohort mock, rebase onto recent master

* feat: Cohort JS draft

* feat: initial commit

* feat: allow cohort to create the certification request payload in callback

* feat: move changes to cohort_v2

* feat: wrap up using a closure get the request + snapshot before sending for certification

* feat: copy cohort_v2 temporary changes for callback to cohort

* feat: Pass the statemaps from the request created by the callback fn

* temp-commit

* feat: use callback for installer

* Integrate OO installer.

* fix: lint and fmt on updating rust to 1.72

* add some stats

* Add tools for bounded concurrency

* feat: Upgrade Pond utility.

* fix: Disable coverage for cohort_sdk_js to see if this fixes build error.

* fix: Disable coverage for cohort_sdk_js to see if this fixes build error.

* fix: Fix conflicts after merging with master

* chore: Fix formatting

* chore: Add index.d.ts to .gitignore

* Remove generated files.

* feat: Rename ItemStateProviderImpl to NewRequestProvider

* feat: Rename cohort_banking_js to cohort_banking_initiator_js to make room for similar replicator app

---------

+++ examples/cohort_replicator_kafka_pg/examples/cohort_replicator_kafka_pg.rs
@@ -59,7 +59,7 @@ async fn main() {
         password: env_var!("COHORT_PG_PASSWORD"),
         port: env_var!("COHORT_PG_PORT"),
         user: env_var!("COHORT_PG_USER"),
-        pool_size: env_var_with_defaults!("COHORT_PG_POOL_SIZE", usize, 10),
+        pool_size: env_var_with_defaults!("COHORT_PG_POOL_SIZE", u32, 10),
     };
     let database = Database::init_db(cfg_db).await.map_err(|e| e.to_string()).unwrap();
