+++ packages/cohort_sdk/src/model/mod.rs
@@ -84,8 +84,8 @@ pub struct Config {
     pub agent: String,
     pub cohort: String,
     // The size of internal buffer for candidates
-    pub buffer_size: usize,
-    pub timeout_ms: u64,
+    pub buffer_size: u32,
+    pub timeout_ms: u32,
 
     //
     // Kafka configs for Agent
